+++ pyccel/codegen/wrapper/c_to_python_wrapper.py
@@ -9,7 +9,7 @@ which creates an interface exposing C code to Python.
 """
 import warnings
 from pyccel.ast.bind_c        import BindCFunctionDef, BindCPointer, BindCFunctionDefArgument
-from pyccel.ast.bind_c        import BindCModule, BindCVariable
+from pyccel.ast.bind_c        import BindCModule, BindCVariable, BindCFunctionDefResult
 from pyccel.ast.builtins      import PythonTuple
 from pyccel.ast.core          import Interface, If, IfSection, Return, FunctionCall
 from pyccel.ast.core          import FunctionDef, FunctionDefArgument, FunctionDefResult
@@ -25,6 +25,7 @@ from pyccel.ast.cwrapper      import PyModule_AddObject, Py_DECREF, PyObject_Typ
 from pyccel.ast.cwrapper      import Py_INCREF, PyType_Ready, WrapperCustomDataType
 from pyccel.ast.c_concepts    import ObjectAddress, PointerCast
 from pyccel.ast.datatypes     import NativeVoid, NativeInteger, CustomDataType, DataTypeFactory
+from pyccel.ast.datatypes     import NativeNumeric
 from pyccel.ast.internals     import get_final_precision
 from pyccel.ast.literals      import Nil, LiteralTrue, LiteralString, LiteralInteger
 from pyccel.ast.numpy_wrapper import pyarray_to_ndarray
@@ -90,7 +91,7 @@ class CToPythonWrapper(Wrapper):
                 var = Variable(dtype=self._python_object_map[dtype],
                                 name=self.scope.get_new_name(name),
                                 memory_handling='alias',
-                                cls_base = self.scope.find(dtype.name, 'classes'),
+                                cls_base = self.scope.find(dtype.name, 'classes', raise_if_missing = True),
                                 is_temp=is_temp)
             except KeyError as e:
                 raise NotImplementedError("Can't return an object whose type was imported. See #1650") from e
@@ -120,7 +121,7 @@ class CToPythonWrapper(Wrapper):
         list of Variable
             Variables which will hold the arguments in Python.
         """
-        collect_args = [self.get_new_PyObject(a.var.name+'_obj', a.var.dtype) for a in args]
+        collect_args = [self.get_new_PyObject(getattr(a, 'original_function_argument_variable', a.var).name+'_obj') for a in args]
         self._python_object_map.update(dict(zip(args, collect_args)))
         return collect_args
 
@@ -245,7 +246,7 @@ class CToPythonWrapper(Wrapper):
         error_code = ()
         dtype = arg.dtype
         if isinstance(dtype, CustomDataType):
-            python_cls_base = self.scope.find(dtype.name, 'classes')
+            python_cls_base = self.scope.find(dtype.name, 'classes', raise_if_missing = True)
             func_call = FunctionCall(PyObject_TypeCheck, [py_obj, python_cls_base.type_object])
         elif rank == 0:
             prec  = arg.precision
@@ -783,20 +784,19 @@ class CToPythonWrapper(Wrapper):
         # This function creates variables so it must be called before extracting them from the scope.
         result_wrap = [l for r in python_results for l in self._wrap(r)]
 
-        # Get the names of the arguments which should be used to call the C-compatible function
-        func_call_arg_names = []
-        for a in original_c_args:
-            if isinstance(a, BindCFunctionDefArgument):
-                orig_var = a.original_function_argument_variable
-                if orig_var.is_optional and not orig_var.is_ndarray:
-                    func_call_arg_names.append(orig_var.name)
-                    continue
-            func_call_arg_names.append(a.var.name)
-
         # Get the arguments and results which should be used to call the c-compatible function
-        func_call_args = [self.scope.find(self.scope.get_expected_name(n), category='variables') for n in func_call_arg_names]
-        c_result_names = [self.scope.get_expected_name(r.var.name) for r in original_c_results]
-        c_results = [self.scope.find(n, category='variables') for n in c_result_names]
+        func_call_args = [self.scope.find(n.var.name, category='variables', raise_if_missing = True) for n in original_c_args]
+
+        # Get the names of the results collected from the C-compatible function
+        c_result_names = []
+        for r in original_c_results:
+            if isinstance(r, BindCFunctionDefResult):
+                orig_var = r.original_function_result_variable
+                if orig_var.rank == 0 and orig_var.dtype in NativeNumeric:
+                    c_result_names.append(orig_var.name)
+                    continue
+            c_result_names.append(r.var.name)
+        c_results = [self.scope.find(n, category='variables', raise_if_missing = True) for n in c_result_names]
         for n, r, o_r in zip(c_result_names, c_results, original_c_results):
             if isinstance(r, DottedVariable):
                 self.scope.remove_variable(r, name=n)
@@ -819,8 +819,8 @@ class CToPythonWrapper(Wrapper):
         # is an ndarray.
         for a in original_c_args:
             orig_var = getattr(a, 'original_function_argument_variable', a.var)
-            v = self.scope.find(self.scope.get_expected_name(orig_var.name), category='variables')
-            if v.is_ndarray:
+            if orig_var.is_ndarray:
+                v = self.scope.find(orig_var.name, category='variables', raise_if_missing = True)
                 if v.is_optional:
                     body.append(If( IfSection(PyccelIsNot(v, Nil()), [Deallocate(v)]) ))
                 else:
@@ -895,10 +895,16 @@ class CToPythonWrapper(Wrapper):
         if orig_var.is_ndarray or isinstance(orig_var.dtype, CustomDataType):
             arg_var = orig_var.clone(self.scope.get_expected_name(orig_var.name), is_argument = False, memory_handling='alias')
             self._wrapping_arrays = orig_var.is_ndarray
+            self.scope.insert_variable(arg_var, orig_var.name)
         else:
-            arg_var = orig_var.clone(self.scope.get_expected_name(orig_var.name), is_argument = False)
+            kwargs = {'is_argument':False}
+            if isinstance(orig_var.dtype, CustomDataType):
+                kwargs['memory_handling']='alias'
+                if isinstance(expr, BindCFunctionDefArgument):
+                    kwargs['dtype'] = NativeVoid()
 
-        self.scope.insert_variable(arg_var)
+            arg_var = orig_var.clone(self.scope.get_expected_name(expr.var.name), **kwargs)
+            self.scope.insert_variable(arg_var, expr.var.name)
 
         body = []
 
@@ -913,9 +919,9 @@ class CToPythonWrapper(Wrapper):
         # Collect the function which casts from a Python object to a C object
         dtype = orig_var.dtype
         if isinstance(dtype, CustomDataType):
-            python_cls_base = self.scope.find(dtype.name, 'classes')
+            python_cls_base = self.scope.find(dtype.name, 'classes', raise_if_missing = True)
             scope = python_cls_base.scope
-            attribute = scope.find('instance', 'variables')
+            attribute = scope.find('instance', 'variables', raise_if_missing = True)
             if expr.bound_argument:
                 cast_type = collect_arg
                 cast = []
@@ -923,13 +929,13 @@ class CToPythonWrapper(Wrapper):
                 cast_type = Variable(dtype=self._python_object_map[dtype],
                                     name=self.scope.get_new_name(collect_arg.name),
                                     memory_handling='alias',
-                                    cls_base = self.scope.find(dtype.name, 'classes'))
+                                    cls_base = self.scope.find(dtype.name, 'classes', raise_if_missing = True))
                 self.scope.insert_variable(cast_type)
                 cast = [AliasAssign(cast_type, PointerCast(collect_arg, cast_type))]
             c_res = attribute.clone(attribute.name, new_class = DottedVariable, lhs = cast_type)
             cast.append(AliasAssign(arg_var, PointerCast(c_res, orig_var)))
         elif arg_var.rank == 0:
-            prec  = get_final_precision(arg_var)
+            prec  = get_final_precision(orig_var)
             try :
                 cast_function = py_to_c_registry[(dtype, prec)]
             except KeyError:
@@ -942,18 +948,18 @@ class CToPythonWrapper(Wrapper):
         else:
             cast = [Assign(arg_var, FunctionCall(pyarray_to_ndarray, [collect_arg]))]
 
-        if arg_var.is_optional:
+        if arg_var.is_optional and not isinstance(dtype, CustomDataType):
             memory_var = self.scope.get_temporary_variable(arg_var, name = arg_var.name + '_memory', is_optional = False)
             cast.insert(0, AliasAssign(arg_var, memory_var))
 
         # Create any necessary type checks and errors
         if expr.has_default:
-            check_func, err = self._get_check_function(collect_arg, arg_var, False)
+            check_func, err = self._get_check_function(collect_arg, orig_var, False)
             body.append(If( IfSection(check_func, cast),
                         IfSection(PyccelIsNot(collect_arg, Py_None), [*err, Return([Nil()])])
                         ))
         elif not in_interface:
-            check_func, err = self._get_check_function(collect_arg, arg_var, True)
+            check_func, err = self._get_check_function(collect_arg, orig_var, True)
             body.append(If( IfSection(check_func, cast),
                         IfSection(LiteralTrue(), [*err, Return([Nil()])])
                         ))
@@ -988,21 +994,22 @@ class CToPythonWrapper(Wrapper):
         orig_var = expr.original_function_argument_variable
 
         if orig_var.rank:
+            bound_var_name = expr.var.name
             # Create variable to hold raw data pointer
-            arg_var = expr.var.clone(self.scope.get_expected_name(expr.var.name), is_argument = False)
+            arg_var = expr.var.clone(self.scope.get_expected_name(bound_var_name), is_argument = False)
             # Create variables for the shapes and strides
             shape_vars = [s.clone(self.scope.get_expected_name(s.name), is_argument = False) for s in expr.shape]
             stride_vars = [s.clone(self.scope.get_expected_name(s.name), is_argument = False) for s in expr.strides]
 
             # Add variables to scope
-            self.scope.insert_variable(arg_var, expr.var.name)
+            self.scope.insert_variable(arg_var, bound_var_name)
             for v,s in zip(shape_vars, expr.shape):
                 self.scope.insert_variable(v,s.name)
             for v,s in zip(stride_vars, expr.strides):
                 self.scope.insert_variable(v,s.name)
 
             # Get the C-compatible variable created in self._wrap_FunctionDefArgument
-            c_arg = self.scope.find(self.scope.get_expected_name(orig_var.name), category='variables')
+            c_arg = self.scope.find(orig_var.name, category='variables', raise_if_missing = True)
 
             # Unpack the C-compatible variable
             body.append(AliasAssign(arg_var, FunctionCall(array_get_data, [c_arg])))
@@ -1054,11 +1061,11 @@ class CToPythonWrapper(Wrapper):
             self._wrapping_arrays = True
         elif isinstance(orig_var.dtype, CustomDataType):
             scope = python_res.cls_base.scope
-            attribute = scope.find('instance', 'variables')
+            attribute = scope.find('instance', 'variables', raise_if_missing = True)
             c_res = attribute.clone(attribute.name, new_class = DottedVariable, lhs = python_res)
         else:
             c_res = orig_var.clone(name, is_argument = False)
-        self.scope.insert_variable(c_res, name)
+        self.scope.insert_variable(c_res, orig_var.name)
 
         # Cast from C to Python
         if not isinstance(orig_var.dtype, CustomDataType):
@@ -1103,34 +1110,37 @@ class CToPythonWrapper(Wrapper):
 
         body = []
 
+        orig_var_name = orig_var.name
+        var_name = expr.var.name
+
         if orig_var.rank:
             # C-compatible result variable
-            c_res = orig_var.clone(self.scope.get_new_name(orig_var.name), is_argument = False, memory_handling='alias')
+            c_res = orig_var.clone(self.scope.get_new_name(orig_var_name), is_argument = False, memory_handling='alias')
             self._wrapping_arrays = True
             # Result of calling the bind-c function
-            arg_var = expr.var.clone(self.scope.get_expected_name(expr.var.name), is_argument = False, memory_handling='alias')
+            arg_var = expr.var.clone(self.scope.get_expected_name(var_name), is_argument = False, memory_handling='alias')
             shape_vars = [s.clone(self.scope.get_expected_name(s.name), is_argument = False) for s in expr.shape]
             # Save so we can find by iterating over func.results
-            self.scope.insert_variable(arg_var, expr.var.name)
+            self.scope.insert_variable(arg_var, var_name)
             for v,s in zip(shape_vars, expr.shape):
                 self.scope.insert_variable(v,s.name)
             # Save so we can find by iterating over func.bind_c_results
-            self.scope.insert_variable(c_res, orig_var.name)
+            self.scope.insert_variable(c_res, orig_var_name)
 
             body.append(Allocate(c_res, shape = shape_vars, order = orig_var.order, status='unallocated'))
             body.append(AliasAssign(DottedVariable(NativeVoid(), 'raw_data', memory_handling = 'alias', lhs=c_res), arg_var))
         elif isinstance(orig_var.dtype, CustomDataType):
-            c_res = expr.var.clone(self.scope.get_expected_name(expr.var.name), is_argument = False, memory_handling='alias')
-            self.scope.insert_variable(c_res, expr.var.name)
+            c_res = expr.var.clone(self.scope.get_expected_name(var_name), is_argument = False, memory_handling='alias')
+            self.scope.insert_variable(c_res, var_name)
         else:
-            c_res = orig_var.clone(self.scope.get_expected_name(orig_var.name), is_argument = False)
-            self.scope.insert_variable(c_res)
+            c_res = orig_var.clone(self.scope.get_expected_name(orig_var_name), is_argument = False)
+            self.scope.insert_variable(c_res, orig_var_name)
 
         if not isinstance(orig_var.dtype, CustomDataType):
             body.append(AliasAssign(python_res, FunctionCall(C_to_Python(c_res), [c_res])))
         else:
             scope = python_res.cls_base.scope
-            attribute = scope.find('instance', 'variables')
+            attribute = scope.find('instance', 'variables', raise_if_missing = True)
             attrib_var = attribute.clone(attribute.name, new_class = DottedVariable, lhs = python_res)
             body.append(AliasAssign(attrib_var, c_res))