+++ pyccel/codegen/wrapper/fortran_to_c_wrapper.py
@@ -97,6 +97,8 @@ class FortranToCWrapper(Wrapper):
                      for fa in func_def_args
                      if not isinstance(func_arg_to_call_arg[fa], IndexedElement) \
                         and fa.original_function_argument_variable.is_optional]
+            body += [C_F_Pointer(fa.var, func_arg_to_call_arg[fa]) for fa in func_def_args
+                    if isinstance(func_arg_to_call_arg[fa].dtype, CustomDataType)]
 
             # If the function is inlined and takes an array argument create a pointer to ensure that the bounds
             # are respected
@@ -132,7 +134,7 @@ class FortranToCWrapper(Wrapper):
             being wrapped.
         """
         original_arg = bind_c_arg.original_function_argument_variable
-        arg_var = self.scope.find(original_arg.name, category='variables')
+        arg_var = self.scope.find(self.scope.get_expected_name(original_arg.name), category='variables')
         if original_arg.is_ndarray:
             start = LiteralInteger(1) # C_F_Pointer leads to default Fortran lbound
             stop = None
@@ -390,7 +392,7 @@ class FortranToCWrapper(Wrapper):
             # Define the additional steps necessary to define and fill ptr_var
             alloc = Allocate(ptr_var, shape=result.shape,
                              order=var.order, status='unallocated')
-            copy = Assign(ptr_var, var)
+            copy = Assign(ptr_var, local_var)
             c_loc = CLocFunc(ptr_var, bind_var)
             self._additional_exprs.extend([alloc, copy, c_loc])