+++ CHANGELOG.md
@@ -5,12 +5,15 @@ All notable changes to this project will be documented in this file.
 
 ### Added
 
+-   #1649 : Add support for `np.min` in C code.
+-   #1621 : Add support for `np.max` in C code.
 -   #1571 : Add support for the function `tuple`.
 -   #1493 : Add preliminary support for importing classes.
 -   #1578 : Allow classes to avoid type annotations for the self argument of a method.
 -   #1597 : Handle class docstrings.
 -   \[INTERNALS\] Add `class_type` attribute to `TypedAstNode`.
 -   #1494 : Add support for functions returning class instances.
+-   #1495 : Add support for functions with class instance arguments.
 
 ### Fixed
 
@@ -25,9 +28,16 @@ All notable changes to this project will be documented in this file.
 -   #1615 : Fixed infinite loop when passing slices while copying arrays.
 -   #1628 : Fixed segmentation fault when writing to optional scalars.
 -   #1554 : Fix exit statement in Fortran with Intel compiler.
+-   #1564 : Fixed installation problems on Python 3.12.
+-   #1259 : Fix bug causing problems with user editable installation.
+-   #1651 : Fix name collision resolution to include parent scopes.
+-   #1156 : Raise an error for variable name collisions with non-variable objects.
 
 ### Changed
 
+-   #1672 : Make `icx` and `ifx` the default Intel compilers (Found in Intel oneAPI).
+-   #1644 : Stop printing the step of a range if that step is 1.
+-   #1638 : Migrate from `setuptools` to `hatch` for installation scripts.
 -   \[INTERNALS\] #1593 : Rename `PyccelAstNode.fst` to the `PyccelAstNode.ast`.
 -   \[INTERNALS\] #1593 : Use a setter instead of a method to update `PyccelAstNode.ast`.
 -   \[INTERNALS\] #1593 : Rename `BasicParser._current_fst_node` to the `BasicParser._current_ast_node`.
@@ -186,7 +196,7 @@ All notable changes to this project will be documented in this file.
 -   #1404 : Bug preventing printing of some functions in a `print()` call.
 -   #1302 : Raise error message in case of empty class.
 -   #1407 : Raise an error if file name matches a Python built-in module.
--   #929 : Allow optional variables when compiling with intel or nvidia.
+-   #929 : Allow optional variables when compiling with Intel or NVIDIA.
 -   #1117 : Allow non-contiguous arrays to be passed to Fortran code.
 -   #1415 : Fix incorrect handling of assignments augmented by function calls.
 -   #1418 : Fix `itertools.product` implementation.